requirements.txt
dj-database-url==0.5.0
Django==3.2.4
gunicorn==20.0.4
PyDictionary==2.0.1
psycopg2-binary==2.8.6
whitenoise==5.2.0
requests==2.25.1
python-decouple==3.4